+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsNetworks.java
@@ -21,10 +21,12 @@ package oshi.hardware.platform.windows;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.sun.jna.platform.win32.WinDef.ULONG;
+
 import oshi.hardware.NetworkIF;
 import oshi.hardware.common.AbstractNetworks;
 import oshi.jna.platform.windows.IPHlpAPI;
-import oshi.jna.platform.windows.IPHlpAPI.MIB_IFROW;
+import oshi.jna.platform.windows.IPHlpAPI.MIB_IFROW2;
 
 /**
  * @author widdis[at]gmail[dot]com
@@ -43,21 +45,19 @@ public class WindowsNetworks extends AbstractNetworks {
      *            The interface on which to update statistics
      */
     public static void updateNetworkStats(NetworkIF netIF) {
-        // Create new MIB_IFROW and set index to this interface index
-        MIB_IFROW ifRow = new MIB_IFROW();
-        ifRow.dwIndex = netIF.getNetworkInterface().getIndex();
-        if (0 != IPHlpAPI.INSTANCE.GetIfEntry(ifRow)) {
+        // Create new MIB_IFROW2 and set index to this interface index
+        MIB_IFROW2 ifRow = new MIB_IFROW2();
+        ifRow.InterfaceIndex = new ULONG(netIF.getNetworkInterface().getIndex());
+        if (0 != IPHlpAPI.INSTANCE.GetIfEntry2(ifRow)) {
             // Error, abort
             LOG.error("Failed to retrieve data for interface {}, {}", netIF.getNetworkInterface().getIndex(),
                     netIF.getName());
             return;
         }
-
-        netIF.setBytesSent(ifRow.dwOutOctets);
-        netIF.setBytesRecv(ifRow.dwInOctets);
-        netIF.setPacketsSent(ifRow.dwOutUcastPkts);
-        netIF.setPacketsRecv(ifRow.dwInUcastPkts);
-        netIF.setSpeed(ifRow.dwSpeed);
+        netIF.setBytesSent(ifRow.OutOctets);
+        netIF.setBytesRecv(ifRow.InOctets);
+        netIF.setPacketsSent(ifRow.OutUcastPkts);
+        netIF.setPacketsRecv(ifRow.InUcastPkts);
+        netIF.setSpeed(ifRow.ReceiveLinkSpeed);
     }
-
 }