+++ oshi-core/src/main/java/oshi/jna/platform/windows/IPHlpAPI.java
@@ -24,6 +24,9 @@ import java.util.List;
 import com.sun.jna.Library;
 import com.sun.jna.Native;
 import com.sun.jna.Structure;
+import com.sun.jna.platform.win32.Guid.GUID;
+import com.sun.jna.platform.win32.WinDef.UCHAR;
+import com.sun.jna.platform.win32.WinDef.ULONG;
 
 /**
  * Windows IP Helper API. This class should be considered non-API as it may be
@@ -34,6 +37,8 @@ import com.sun.jna.Structure;
 public interface IPHlpAPI extends Library {
     IPHlpAPI INSTANCE = (IPHlpAPI) Native.loadLibrary("IPHlpAPI", IPHlpAPI.class);
 
+    int IF_MAX_STRING_SIZE = 256;
+    int IF_MAX_PHYS_ADDRESS_LENGTH = 32;
     int MAX_INTERFACE_NAME_LEN = 256;
     int MAXLEN_IFDESCR = 256;
     int MAXLEN_PHYSADDR = 8;
@@ -74,6 +79,65 @@ public interface IPHlpAPI extends Library {
         }
     }
 
+    class MIB_IFROW2 extends Structure {
+        public long InterfaceLuid; // 64-bit union
+        public ULONG InterfaceIndex;
+        public GUID InterfaceGuid;
+        public char[] Alias = new char[IF_MAX_STRING_SIZE + 1];
+        public char[] Description = new char[IF_MAX_STRING_SIZE + 1];
+        public ULONG PhysicalAddressLength;
+        public UCHAR[] PhysicalAddress = new UCHAR[IF_MAX_PHYS_ADDRESS_LENGTH];
+        public UCHAR[] PermanentPhysicalAddress = new UCHAR[IF_MAX_PHYS_ADDRESS_LENGTH];
+        public ULONG Mtu;
+        public ULONG Type;
+        // enums
+        public int TunnelType;
+        public int MediaType;
+        public int PhysicalMediumType;
+        public int AccessType;
+        public int DirectionType;
+        // 8-bit structure
+        public byte InterfaceAndOperStatusFlags;
+        // enums
+        public int OperStatus;
+        public int AdminStatus;
+        public int MediaConnectState;
+        public GUID NetworkGuid;
+        public int ConnectionType;
+        public long TransmitLinkSpeed;
+        public long ReceiveLinkSpeed;
+        public long InOctets;
+        public long InUcastPkts;
+        public long InNUcastPkts;
+        public long InDiscards;
+        public long InErrors;
+        public long InUnknownProtos;
+        public long InUcastOctets;
+        public long InMulticastOctets;
+        public long InBroadcastOctets;
+        public long OutOctets;
+        public long OutUcastPkts;
+        public long OutNUcastPkts;
+        public long OutDiscards;
+        public long OutErrors;
+        public long OutUcastOctets;
+        public long OutMulticastOctets;
+        public long OutBroadcastOctets;
+        public long OutQLen;
+
+        @Override
+        protected List<String> getFieldOrder() {
+            return Arrays.asList(new String[] { "InterfaceLuid", "InterfaceIndex", "InterfaceGuid", "Alias",
+                    "Description", "PhysicalAddressLength", "PhysicalAddress", "PermanentPhysicalAddress", "Mtu",
+                    "Type", "TunnelType", "MediaType", "PhysicalMediumType", "AccessType", "DirectionType",
+                    "InterfaceAndOperStatusFlags", "OperStatus", "AdminStatus", "MediaConnectState", "NetworkGuid",
+                    "ConnectionType", "TransmitLinkSpeed", "ReceiveLinkSpeed", "InOctets", "InUcastPkts",
+                    "InNUcastPkts", "InDiscards", "InErrors", "InUnknownProtos", "InUcastOctets", "InMulticastOctets",
+                    "InBroadcastOctets", "OutOctets", "OutUcastPkts", "OutNUcastPkts", "OutDiscards", "OutErrors",
+                    "OutUcastOctets", "OutMulticastOctets", "OutBroadcastOctets", "OutQLen" });
+        }
+    }
+
     /**
      * The GetIfEntry function retrieves information for the specified interface
      * on the local computer
@@ -87,4 +151,16 @@ public interface IPHlpAPI extends Library {
      */
     int GetIfEntry(MIB_IFROW pIfRow);
 
+    /**
+     * The GetIfEntry function retrieves information for the specified interface
+     * on the local computer
+     * 
+     * @param pIfRow
+     *            A pointer to a MIB_IFROW structure that, on successful return,
+     *            receives information for an interface on the local computer.
+     *            On input, set the dwIndex member of MIB_IFROW to the index of
+     *            the interface for which to retrieve information.
+     * @return If the function succeeds, the return value is NO_ERROR.
+     */
+    int GetIfEntry2(MIB_IFROW2 pIfRow2);
 }