+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 3.2 (in progress)
 ================
+* [#243](https://github.com/dblock/oshi/pull/242): Make Windows network statistics 64-bit - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here.
 
 3.1.1 (8/5/2016)